everything is still burning
