--- app/prattle/browser/config/index.js
@@ -1,7 +1,0 @@
-'use strict';
-
-var fs = require('fs');
-
-require('./create-dot-prattle.js');
-
-module.exports = require('./settings-methods.js');
