--- app/prattle/browser/irc/client.js
@@ -1,89 +1,0 @@
-'use strict';
-var irc = require('irc');
-var ipc = require('ipc');
-
-module.exports = function (app, mainWindow, clientData) {
-  // console.log(clientData);
-  mainWindow.client = new irc.Client(clientData.hostAddr, clientData.nickName, {
-    userName: clientData.nickName,
-    realName: clientData.realName,
-    port: 6667
-  });
-
-  /*
-   * Here we have to pretty much re-implement the API for the Client object.
-   * It's very bad for the render process to send callbacks to the main process,
-   *  SEE: https://github.com/atom/electron/blob/master/docs/api/remote.md#passing-callbacks-to-the-main-process
-   * so the rendering process can bind to an event on the mainWindow,
-   * which isn't the main process. This is much cleaner.
-   *
-   * client -> main process -> renderer
-   */
-
-  /* You can find the events at:
-   * http://node-irc.readthedocs.org/en/latest/API.html#events
-   **/
-
-  // TODO: Bind to all the events in a prettier way, instead of writing them out individually
-
-  app.on('before-quit', function (event) {
-    mainWindow.client.disconnect();
-  });
-
-  mainWindow.client.addListener('registered', function (message) {
-    mainWindow.send('connect-ready', message);
-    // console.log(mainWindow.client);
-  });
-
-  mainWindow.client.addListener('error', function (message) {
-    mainWindow.send('client-error', message);
-  });
-
-  mainWindow.client.addListener('motd', function (motd) {
-    mainWindow.send('client-motd', motd);
-  });
-
-  mainWindow.client.addListener('message', function (nick, to, text, message) {
-    mainWindow.send('client-message', nick, to, text, message);
-  });
-
-  mainWindow.client.addListener('selfMessage', function (to, text) {
-    mainWindow.send('client-selfMessage', to, text);
-  });
-
-  mainWindow.client.addListener('names', function (channel, nicks) {
-    mainWindow.send('client-names', nicks);
-  });
-
-  mainWindow.client.addListener('topic', function (channel, topic, nick, message) {
-    mainWindow.send('client-topic', channel, topic, nick, message);
-  });
-
-  mainWindow.client.addListener('kill', function (nick, reason, channels, message) {
-    mainWindow.send('client-kill', nick, reason, channels, message);
-  });
-
-  mainWindow.client.addListener('pm', function (nick, text, message) {
-    mainWindow.send('client-pm', nick, text, message);
-  });
-
-  mainWindow.client.addListener('kick', function (channel, nick, by, reason, message) {
-    mainWindow.send('client-kick', channel, nick, by, reason, message);
-  });
-
-  mainWindow.client.addListener('notice', function (nick, to, text, message) {
-    mainWindow.send('client-notice', nick, to, text, message);
-  });
-
-  mainWindow.client.addListener('action', function (from, to, text, message) {
-    mainWindow.send('client-action', from, to, text, message);
-  });
-
-  mainWindow.client.addListener('raw', function (message) {
-    mainWindow.send('client-raw', message);
-  });
-
-  mainWindow.client.addListener('ctcp', function (from, to, text, type, message) {
-    mainWindow.send('client-ctcp', from, to, text, type, message);
-  });
-};
