+++ app/prattle/browser/menu.js
